+++ database/migrations/2024_12_22_130629_add_jumlah_to_produks_table.php
@@ -12,7 +12,10 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('produks', function (Blueprint $table) {
-            $table->decimal('jumlah', 8, 2)->after('produk_id')->default(0); // Tambahkan kolom jumlah
+            // Periksa apakah kolom 'jumlah' sudah ada sebelum menambahkannya
+            if (!Schema::hasColumn('produks', 'jumlah')) {
+                $table->decimal('jumlah', 8, 2)->after('produk_id')->default(0); // Tambahkan kolom jumlah
+            }
         });
     }
 
@@ -22,7 +25,10 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('produks', function (Blueprint $table) {
-            $table->dropColumn('jumlah');
+            // Hapus kolom 'jumlah' jika ada
+            if (Schema::hasColumn('produks', 'jumlah')) {
+                $table->dropColumn('jumlah');
+            }
         });
     }
 };
